requirements.txt
cycler==0.10.0
decorator==4.2.1
kiwisolver==1.0.1
matplotlib==2.2.0rc1
networkx==2.1
numpy==1.14.0+mkl
numpy-stl==2.4.1
opencv-python==3.4.0
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.3.0
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
setuptools-scm==1.15.7
six==1.11.0
